WIP Split up TrustfallPlayground and Hackernews logic
