+++ experiments/browser_based_querying/src/App.tsx
@@ -1,0 +1,5 @@
+import HackerNews from './playgrounds/HackerNews';
+
+export default function App() {
+  return <HackerNews />;
+}
