+++ experiments/browser_based_querying/src/index.tsx
@@ -2,11 +2,11 @@ import 'core-js/stable';
 import { render } from 'react-dom';
 import { createRoot } from 'react-dom/client';
 
-import HackerNews from './playgrounds/HackerNews';
+import App from './App';
 
 const mainEl = document.getElementsByTagName('main')[0];
 const root = createRoot(mainEl);
-root.render(<HackerNews />);
+root.render(<App />);
 if (process.env.NODE_ENV === 'development' && module.hot) {
-  module.hot.accept('./App', () => render(<HackerNews />, mainEl));
+  module.hot.accept('./App', () => render(<App />, mainEl));
 }
